Upgrade to Python 3.11 + NumPy 2.x, bump version to 1.2.0
- Fix NumPy 2.x compatibility: cast FFT size `n` to int in fftkernel
  and fftkernelWin (np.fft.ifft no longer accepts float for n)
- Add python_requires>=3.9 and install_requires=[numpy>=1.24] to setup.py
- Sync __version__ in __init__.py to 1.2.0 (was 1.0.0)
- Add benchmark runner (tests/run_benchmarks.py) with JSON output
- Record baseline benchmarks: Python 3.9/numpy 1.26 vs 3.11/numpy 2.4

All 21 pytest tests and 18/18 golden checks pass on both envs.

+++ adaptivekde/sskernel.py
@@ -175,11 +175,11 @@ def fftkernel(x, w):
 
     L = x.size
     Lmax = L + 3 * w
-    n = 2 ** np.ceil(np.log2(Lmax))
+    n = int(2 ** np.ceil(np.log2(Lmax)))
 
-    X = np.fft.fft(x, n.astype(int))
+    X = np.fft.fft(x, n)
 
-    f = np.linspace(0, n-1, n.astype(int)) / n
+    f = np.linspace(0, n-1, n) / n
     f = np.concatenate((-f[0: int(n / 2 + 1)],
                         f[1: int(n / 2 - 1 + 1)][::-1]))
